06/04/2017 DM - Busqueda de archivo en la BD

+++ ckeckingdocuments/src/main/java/com/checkingdocuments/service/ReunionesOperativasService.java
@@ -6,11 +6,11 @@ import com.checkingdocuments.model.ReunionesOperativas;
 
 public interface ReunionesOperativasService {
 	
-	public ReunionesOperativas findById(Integer id);
+	public ReunionesOperativas findById(Long id);
 	public List<ReunionesOperativas> findByResponsable(String name);
 	public void saveReunionesOperativas(ReunionesOperativas reunionesOperativas);
 	public void updateReunionesOperativas(ReunionesOperativas reunionesOperativas);
-	public void deleteReunionesOperativasById(Integer id);
+	public void deleteReunionesOperativasById(Long id);
 	public List<ReunionesOperativas> findAllReunionesOperativas();
 	
 }
